Self Project/v2/requirements.txt
Flask==3.1.0
line-bot-sdk
azure-storage-blob
azure-cognitiveservices-speech
azure-ai-textanalytics
azure-core
google-generativeai
pillow
google-cloud-firestore
requests